feat(qa): warn when ae.config coverage thresholds present; policy is source of truth (#918)
