+++ artifacts/domain/replay.summary.json
@@ -1,0 +1,1 @@
+{"traceId":"t-1","totalEvents":3,"violatedInvariants":[]}
