+++ docs/contributing/PR_TEMPLATE.md
@@ -64,9 +64,10 @@ export default {
 ```
 
 Note
+- Precedence: policy > AE-IR > ae.config
 - The values under `qa.coverageThreshold` in `ae.config.ts` are treated as hints for local visibility.
 - Enforcement uses the centralized policy at `policy/quality.json` (profile via `AE_QUALITY_PROFILE` or `CI` → `ci`).
-- If `ae.config.ts` differs from policy, the QA CLI prints a warning and uses the policy thresholds as the source of truth.
+- When `ae.config.ts` contains thresholds, the QA CLI prints a warning and uses the policy thresholds as the source of truth.
 
 ## ✅ 検証結果
