+++ reports/benchmark/req2run-benchmark-2025-09-19T17-13-06-613Z.md
@@ -1,0 +1,13 @@
+# Req2Run Benchmark Report
+
+Generated: 2025-09-19T17:13:06.614Z
+
+## Summary
+- **Total Problems**: 0
+- **Successful Runs**: 0
+- **Failed Runs**: 0
+- **Average Score**: 0.00/100
+- **Total Execution Time**: 0ms
+
+## Individual Results
+
