+++ reports/benchmark/req2run-benchmark-2025-09-19T17-13-07-799Z.md
@@ -1,0 +1,23 @@
+# Req2Run Benchmark Report
+
+Generated: 2025-09-19T17:13:07.799Z
+
+## Summary
+- **Total Problems**: 2
+- **Successful Runs**: 0
+- **Failed Runs**: 2
+- **Average Score**: 0.00/100
+- **Total Execution Time**: 5ms
+
+## Individual Results
+### problem-1
+- **Status**: ❌ Failed
+- **Score**: 0/100
+- **Execution Time**: 2ms
+- **Errors**: Failed to load problem spec for problem-1: Problem specification not found for problem-1
+
+### problem-2
+- **Status**: ❌ Failed
+- **Score**: 0/100
+- **Execution Time**: 3ms
+- **Errors**: Failed to load problem spec for problem-2: Problem specification not found for problem-2
