+++ src/commands/qa/run.ts
@@ -1,7 +1,7 @@
 import { execa } from 'execa';
-import { loadConfig, type AeConfig } from '../../core/config.js';
+import { loadConfig } from '../../core/config.js';
 import { readFile, stat } from 'node:fs/promises';
-import { getQualityGate } from '../../utils/quality-policy-loader.js';
+import { resolveCoverageThresholds } from '../../utils/coverage-thresholds.js';
 
 export async function qaRun(options?: { light?: boolean }) {
   const cfg = await loadConfig();
@@ -15,13 +15,13 @@ export async function qaRun(options?: { light?: boolean }) {
   console.log(`[ae:qa] Using coverage thresholds from policy/quality.json (profile: ${envProfile})`);
   console.log(`[ae:qa] thresholds → lines=${effective.lines}, functions=${effective.functions}, branches=${effective.branches}, statements=${effective.statements}`);
   if (hint) {
+    // Always warn when ae.config contains thresholds; policy remains the source of truth.
+    console.warn('[ae:qa] WARN: ae.config.qa.coverageThreshold is treated as a hint. Policy is the source of truth.');
+    console.warn(`[ae:qa] hint → lines=${hint.lines}, functions=${hint.functions}, branches=${hint.branches}, statements=${hint.statements}`);
     if (mismatch) {
-      console.warn('[ae:qa] WARN: ae.config.qa.coverageThreshold is treated as a hint and differs from policy. Policy is the source of truth.');
-      console.warn(`[ae:qa] hint → lines=${hint.lines}, functions=${hint.functions}, branches=${hint.branches}, statements=${hint.statements}`);
-      console.warn('[ae:qa] To change enforcement, update policy/quality.json (coverage.thresholds.*) or set AE_QUALITY_PROFILE.');
-    } else {
-      console.log('[ae:qa] Note: ae.config.qa.coverageThreshold is treated as a hint; policy governs enforcement.');
+      console.warn('[ae:qa] HINT differs from policy thresholds. Enforcement will follow policy.');
     }
+    console.warn('[ae:qa] To change enforcement, update policy/quality.json (coverage.thresholds.*) or set AE_QUALITY_PROFILE.');
   }
   
   if (runner === 'jest') {
@@ -58,45 +58,7 @@ export async function qaRun(options?: { light?: boolean }) {
   console.log('[ae:qa] QA checks completed');
 }
 
-export type CoverageThresholds = { branches: number; lines: number; functions: number; statements: number };
-
-/**
- * Resolve effective coverage thresholds from centralized policy, treating ae.config as a hint.
- * Falls back to ae.config when policy load fails.
- */
-export async function resolveCoverageThresholds(cfg: AeConfig, envProfile: string): Promise<{
-  effective: CoverageThresholds;
-  hint: CoverageThresholds | null;
-  mismatch: boolean;
-}> {
-  const hint = cfg?.qa?.coverageThreshold ?? null;
-  try {
-    const gate = getQualityGate('coverage', envProfile);
-    const eff: CoverageThresholds = {
-      lines: Number(gate.thresholds.lines ?? 80),
-      functions: Number(gate.thresholds.functions ?? 80),
-      branches: Number(gate.thresholds.branches ?? 80),
-      statements: Number(gate.thresholds.statements ?? 80),
-    };
-    const mismatch = hint ? (
-      hint.lines !== eff.lines ||
-      hint.functions !== eff.functions ||
-      hint.branches !== eff.branches ||
-      hint.statements !== eff.statements
-    ) : false;
-    return { effective: eff, hint, mismatch };
-  } catch (e) {
-    // Policy load failed — fall back to config hint with a warning
-    if (hint) {
-      console.warn('[ae:qa] WARN: Failed to load policy/quality.json. Falling back to ae.config.qa.coverageThreshold.');
-      return { effective: hint, hint, mismatch: false };
-    }
-    // Final fallback
-    const eff: CoverageThresholds = { lines: 80, functions: 80, branches: 80, statements: 80 };
-    console.warn('[ae:qa] WARN: No policy and no ae.config thresholds found. Falling back to defaults (80%).');
-    return { effective: eff, hint: null, mismatch: false };
-  }
-}
+// resolveCoverageThresholds moved to utils for testability
 
 async function detectPM(): Promise<'pnpm' | 'npm' | 'yarn' | null> {
   try { await stat('pnpm-lock.yaml'); return 'pnpm'; } catch {}
