+++ tests/golden/snapshots/codegen-snapshot.json
@@ -1,5 +1,5 @@
 {
-  "timestamp": "2025-09-18T07:17:11.813Z",
+  "timestamp": "2025-09-19T17:13:04.378Z",
   "version": "1.0.0",
   "files": {
     "examples/inventory/apps/web/components/ProductForm.tsx": {
@@ -93,18 +93,18 @@
       "typeScriptErrors": 1,
       "eslintErrors": 0
     },
-    "examples/inventory/apps/web/app/customer/page.tsx": {
-      "hash": "9889197ef585928b8a3d03a201f2df7d1dec53af82d475b495e1b48354b0011c",
-      "lineCount": 108,
-      "ariaAttributeCount": 0,
-      "typeScriptErrors": 0,
-      "eslintErrors": 0
-    },
     "examples/inventory/apps/web/app/order/page.tsx": {
       "hash": "9b1dc2fa35a32648d7c4ec5186df715a8c6d37b7ebe17aed14fe2fda193bb9bf",
       "lineCount": 118,
       "ariaAttributeCount": 0,
       "typeScriptErrors": 1,
+      "eslintErrors": 0
+    },
+    "examples/inventory/apps/web/app/customer/page.tsx": {
+      "hash": "9889197ef585928b8a3d03a201f2df7d1dec53af82d475b495e1b48354b0011c",
+      "lineCount": 108,
+      "ariaAttributeCount": 0,
+      "typeScriptErrors": 0,
       "eslintErrors": 0
     },
     "examples/inventory/apps/web/app/product/new/page.tsx": {
@@ -121,20 +121,6 @@
       "typeScriptErrors": 1,
       "eslintErrors": 0
     },
-    "examples/inventory/apps/web/app/customer/new/page.tsx": {
-      "hash": "ddaf62479aa61e9eaa97d021c29c05d438bf740c1964a91f75eada1da6ed6263",
-      "lineCount": 75,
-      "ariaAttributeCount": 0,
-      "typeScriptErrors": 0,
-      "eslintErrors": 0
-    },
-    "examples/inventory/apps/web/app/customer/[id]/page.tsx": {
-      "hash": "2a001e47fd626011516478ac7f1f6930ee748e6e68a37c02a3c3c68c98e4c5e4",
-      "lineCount": 214,
-      "ariaAttributeCount": 0,
-      "typeScriptErrors": 0,
-      "eslintErrors": 0
-    },
     "examples/inventory/apps/web/app/order/new/page.tsx": {
       "hash": "b62a69782240c7feb695b07f38478fdf9585bb68b7f6375ec4e6a02fc0b9263e",
       "lineCount": 85,
@@ -147,6 +133,20 @@
       "lineCount": 305,
       "ariaAttributeCount": 0,
       "typeScriptErrors": 1,
+      "eslintErrors": 0
+    },
+    "examples/inventory/apps/web/app/customer/new/page.tsx": {
+      "hash": "ddaf62479aa61e9eaa97d021c29c05d438bf740c1964a91f75eada1da6ed6263",
+      "lineCount": 75,
+      "ariaAttributeCount": 0,
+      "typeScriptErrors": 0,
+      "eslintErrors": 0
+    },
+    "examples/inventory/apps/web/app/customer/[id]/page.tsx": {
+      "hash": "2a001e47fd626011516478ac7f1f6930ee748e6e68a37c02a3c3c68c98e4c5e4",
+      "lineCount": 214,
+      "ariaAttributeCount": 0,
+      "typeScriptErrors": 0,
       "eslintErrors": 0
     },
     "examples/inventory/src/ui/components/generated/apps/web/components/ProductForm.tsx": {
